Make a deep item to dict conversion on ItemValidationPipeline
Instead of just using dict(item) now we use JsonLinesItemExporter to be
able to have a deep conversion.

This allows us to validate nested fields in a realiable way.

We were getting wrong validation errors because we have items with nested fields
that are also items, which don't happen with this patch.