Rename setting variable to only add field coverage info in stats
The initial idea was to have a setting to contain the field coverage rules
that would be used by a built-in monitor that will check field coverage
results.

With this new setting name, user will be able to just inform if she wants
to add the field coverage statistics in spider stats and decide later
if the built-in monitor will be used (by defining the field coverage dictionary)
or a custom monitor will be used (and having access to these stats).